requirements.txt
pylint
pytest
click
black
pytest-cov

python-dotenv
databricks-sql-connector
matplotlib

pyspark
